requirements.txt
# Core Framework
sentient-agent-framework==0.3.0

# Web Framework
fastapi==0.115.12
uvicorn==0.34.0
starlette==0.46.1

# Blockchain Dependencies
web3==6.15.1

# Search Dependencies  
tavily-python==0.5.4

# HTTP Client
httpx==0.28.1
requests==2.32.3

# Data Processing
pydantic==2.11.3
pydantic_core==2.33.1

# Environment Management
python-dotenv==1.1.0

# Utilities
python-ulid==3.0.0
cuid2==2.0.1
PyYAML==6.0.2
tenacity==9.1.2
tqdm==4.67.1

# Standard Library Extensions
typing_extensions==4.13.2
typing-inspection==0.4.0
annotated-types==0.7.0

# HTTP and Network
anyio==4.9.0
sniffio==1.3.1
h11==0.14.0
httpcore==1.0.7
idna==3.10
certifi==2025.1.31
charset-normalizer==3.4.1
urllib3==2.4.0

# JSON Processing
jsonpatch==1.33
jsonpointer==3.0.0

# CLI
click==8.1.8

# Compression
zstandard==0.23.0

# Package Management
packaging==24.2
regex==2024.11.6
requests-toolbelt==1.0.0

# Async Support
jiter==0.9.0
distro==1.9.0
